split to three separate modules - cloudfront_distribution, cloudfront_invalidation and cloudfront_origin_access_identity
